avc memoire + projet PermApp

+++ JAVA EE/SpringEtORM/permApp/src/main/java/com/permApp/service/PlanteService.java
@@ -18,7 +18,7 @@ public class PlanteService {
 		return planteDao.getPlantes();
 	}
 
-	public Plante getPlante(long planteId) {
+	public Plante getPlante(int planteId) {
 		return planteDao.getPlante(planteId);
 	}
